# docs/swagger.yaml
basePath: /
definitions:
...
  handler.ExistingFile:
    type: object
...
  handler.File2View:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      type:
        type: string
      updated:
        type: string
      url:
        type: string
    type: object
...
  handler.FilesListItem:
    properties:
      created:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      type:
        type: string
      updated:
        type: string
      url:
        type: string
    type: object
  handler.FilesListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handler.FilesListItem'
        type: array
      filtered:
        type: integer
      total:
        type: integer
    type: object
...
  handler.HTTPError:
    properties:
      error:
        type: string
    type: object
  handler.HTTPMessage:
    properties:
      message:
        type: string
    type: object
...
  handler.ListRequest:
    properties:
      filter:
        additionalProperties:
          type: string
        type: object
      length:
        type: integer
      sort:
        additionalProperties:
          type: string
        type: object
      start:
        type: integer
    type: object
...
  handler.NewFile:
    properties:
      file:
        type: string
      name:
        type: string
    type: object
...
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: GoShop full featured api, see documentation https://docs.google.com/document/d/1VlkAYTqZG9oGvZxnSNDpcuafU-msOb2Qm2FevMJkA7w/edit
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: "1.0"
paths:
...
  /api/v1/files:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: body
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/handler.NewFile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Create file
      tags:
      - file
  /api/v1/files/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Delete file
      tags:
      - file
    get:
      consumes:
      - application/json
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.File2View'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Get file
      tags:
      - file
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/handler.ExistingFile'
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Update file
      tags:
      - file
  /api/v1/files/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.FilesListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Search files
      tags:
      - file
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"